requirements.txt
python-dotenv
sounddevice
numpy
httpx
torch
torchaudio
silero-vad